morphs/ladimo/requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
albucore==0.0.17
albumentations==1.4.18
altair==5.4.1
annotated-types==0.7.0
antlr4-python3-runtime==4.8
astor==0.8.1
async-timeout==4.0.3
attrs==24.2.0
backports.zoneinfo==0.2.1
base58==2.1.1
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
diffusers==0.30.3
einops==0.3.0
eval_type_backport==0.2.0
filelock==3.16.1
frozenlist==1.4.1
fsspec==2024.9.0
ftfy==6.2.3
future==1.0.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.35.0
google-auth-oauthlib==1.0.0
grpcio==1.66.2
huggingface-hub==0.25.2
idna==3.10
imageio==2.35.1
imageio-ffmpeg==0.4.2
importlib_metadata==8.5.0
importlib_resources==6.4.5
invisible-watermark==0.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kornia==0.6.0
lazy_loader==0.4
lightning-utilities==0.11.7
Markdown==3.7
MarkupSafe==2.1.5
multidict==6.1.0
narwhals==1.9.3
networkx==3.1
numpy==1.24.4
oauthlib==3.2.2
omegaconf==2.1.1
opencv-python==4.7.0.72
opencv-python-headless==4.10.0.84
packaging==24.1
pandas==2.0.3
pillow==10.4.0
pkgutil_resolve_name==1.3.10
propcache==0.2.0
protobuf==5.28.2
pudb==2019.2
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
pyDeprecate==0.3.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
pytorch-lightning==1.4.2
pytz==2024.2
PyWavelets==1.4.1
PyYAML==6.0.2
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.20.0
rsa==4.9
safetensors==0.4.5
scikit-image==0.21.0
scipy==1.10.1
six==1.16.0
smmap==5.0.1
streamlit==0.73.1
tensorboard==2.14.0
tensorboard-data-server==0.7.2
test_tube==0.7.5
tifffile==2023.7.10
tokenizers==0.12.1
toml==0.10.2
torch==1.11.0+cu113
torch-fidelity==0.3.0
torchaudio==0.11.0+cu113
torchmetrics==0.6.0
torchvision==0.12.0+cu113
tornado==6.4.1
tqdm==4.66.5
transformers==4.19.2
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
urwid==2.6.15
validators==0.34.0
watchdog==4.0.2
wcwidth==0.2.13
Werkzeug==3.0.4
yarl==1.15.0
zipp==3.20.2
